discord-engine/requirements.txt
discord.py==2.4.0
openai==1.57.0
google-generativeai==0.8.1
anthropic==0.40.0
python-dotenv==1.0.1
aiohttp==3.10.10
httpx==0.27.2
